⚡ Bolt: Extract chat input to prevent O(N) list re-renders
Extracted the input form into its own memoized component (`ChatInputForm.tsx`).

This significantly improves performance for the ChatPanel when there are many messages by isolating the input state. Previously, typing triggered the entire parent `ChatPanel` component to re-render on every keystroke, which evaluates all historical messages. Now, only the isolated input field re-renders during typing.
